Alts registration per server (#248)
* feat: now libertybans should support alt registration per server.

This means that now you can choose whether all servers will register alts or only the selected ones. For proxies, you need to provide the names of the servers that you have in your configs, and for backends just enable or disable a setting in config.

* refactor

* feat: should be working now [not tested yet]

* fix: moved check outside of db query.

* feat: added more precise explanation

* Accomodate unrecorded IP address in applicability calculation

Plus a couple tweaks to code style

* refactoring

* refactoring

* refactor: config typo

* refactor: typos

* fix: remove usage of join()

* Revert "fix: remove usage of join()"

This reverts commit 3e499807938c8d3d0b3d22c884d2ecfbd70d656d.

* Add integration tests + skip canceled server switch on bungee

* Set maven-invoker to output error logs to console

* Document the feature on the wiki

---------

Co-authored-by: A248 <a248-dev@tuta.io>